Bump checkout GHA to v3 from v2
The checkout GitHub Action was using v2 which is deprecated due to use
of Node.js 12 in favour of Node.js 16. The checkout v2 usage in our
workflow definitions isn't using any parameters that aren't compatible
with v3, so no changes other than the version bump is expected to be
needed here.