Login to the infrawatch quay registry
Created organization secrets in GitHub with the
infrawatch+release_automation robot account so that the release
automation script (for nightly builds) can now push to the infrawatch
registry.

Previously the release automation script only needed to access the
infrawatch-operators registry to push the bundles and index image, but
with the recent changes to tag and push the latest into the nightly
builds, access to the infrawatch repo is also necessary.

Related: STF-1559